requirements.txt
asgiref==3.5.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
install==1.3.5
Markdown==3.3.6
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.2
whitenoise==6.2.0
